trying to get music library

+++ BTLE_TransferNext/BTLE Transfer/BTLECentralViewController.m
@@ -292,7 +292,7 @@
     }
     
     NSString *stringFromData = [[NSString alloc] initWithData:characteristic.value encoding:NSUTF8StringEncoding];
-    NSLog(@"data %i", self.data.length);
+    NSLog(@"data %lu", (unsigned long)self.data.length);
     // Have we got everything we need?
     if ([stringFromData isEqualToString:@"EOM"]) {
         NSLog(@"characteristic updated");
